Add configmaps read permission for istio-reader clusterrole.
In the design [1], in order to generate correct secure naming, SPIFFE Trustdomain,Istiod needs to figure out the remote cluster's trust
domain, which is stored in MeshConfig as config map. This PR is to allow Istiod with such permission.

[1] https://docs.google.com/document/d/1LHXcWJK0zZzZ7bn1errnS8uUGTxMvqjG9g_yqPv3jMU/edit

This should be all required changes for the manifest:

1. pilot envioronment variable regarding the trust domain verification
can be done in values.istiod.env
2. meshconfig trust domain alias is also already exposed and editable
in profile
3. clusterrole, this one.
4. image update, auto resolved after binary update.

Change-Id: I6ab96e7c08bc88a56d35b238b1b366b733911ad0